Organized classes into individual packages
