+++ spring-cloud-gateway-server-mvc/src/main/java/org/springframework/cloud/gateway/server/mvc/config/FilterProperties.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.server.mvc;
+package org.springframework.cloud.gateway.server.mvc.config;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import jakarta.validation.ValidationException;
 import jakarta.validation.constraints.NotNull;
 
+import org.springframework.cloud.gateway.server.mvc.common.NameUtils;
 import org.springframework.validation.annotation.Validated;
 
 import static org.springframework.util.StringUtils.tokenizeToStringArray;
@@ -31,21 +31,21 @@ import static org.springframework.util.StringUtils.tokenizeToStringArray;
  * @author Spencer Gibb
  */
 @Validated
-public class PredicateProperties {
+public class FilterProperties {
 
 	@NotNull
 	private String name;
 
 	private Map<String, String> args = new LinkedHashMap<>();
 
-	public PredicateProperties() {
+	public FilterProperties() {
 	}
 
-	public PredicateProperties(String text) {
+	public FilterProperties(String text) {
 		int eqIdx = text.indexOf('=');
 		if (eqIdx <= 0) {
-			throw new ValidationException(
-					"Unable to parse PredicateDefinition text '" + text + "'" + ", must be of the form name=value");
+			setName(text);
+			return;
 		}
 		setName(text.substring(0, eqIdx));
 
@@ -84,7 +84,7 @@ public class PredicateProperties {
 		if (o == null || getClass() != o.getClass()) {
 			return false;
 		}
-		PredicateProperties that = (PredicateProperties) o;
+		FilterProperties that = (FilterProperties) o;
 		return Objects.equals(name, that.name) && Objects.equals(args, that.args);
 	}
 
@@ -95,7 +95,7 @@ public class PredicateProperties {
 
 	@Override
 	public String toString() {
-		final StringBuilder sb = new StringBuilder("PredicateDefinition{");
+		final StringBuilder sb = new StringBuilder("FilterDefinition{");
 		sb.append("name='").append(name).append('\'');
 		sb.append(", args=").append(args);
 		sb.append('}');
